fastapi/requirements.txt
aiofiles
annotated-types
anyio
bcrypt
click
datetime
ecdsa
fastapi
h11
idna
passlib
pyasn1
pydantic
pydantic_core
python-jose
python-multipart
rsa
six
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
python-dotenv

plaid-python
